requirements.txt
langchain
ipykernel
openai
fastapi
langchain_groq
langchain_core
langserve
python-dotenv
streamlit
sse_starlette
